requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
eight==1.0.1
future==0.18.2
idna==2.10
lxml==4.6.3
Pillow==8.2.0
pycparser==2.20
PyNFe==0.4
pyOpenSSL==19.1.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
signxml==2.8.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
